+++ R/RcppExports.R
@@ -12,7 +12,6 @@
 #' \itemize{
 #' \item{Data is not compressed.}
 #' \item{Matrix or array must be of type \code{numeric}}
-#' \item{When converting to unsigned bytes for output, values are truncated rather than rounded}
 #' }
 #'
 #' Design decisions