+++ R/write_png.R
@@ -11,7 +11,6 @@
 #' \itemize{
 #' \item{Data is not compressed.}
 #' \item{Matrix or array must be of type \code{numeric}}
-#' \item{When converting to unsigned bytes for output, values are truncated rather than rounded}
 #' }
 #'
 #' @param data numeric 2d matrix or 3d array (with 3 planes)