+++ tests/testthat/test-write-png-against-png-package.R
@@ -42,18 +42,13 @@ test_that("write_png_core works for grey", {
       system(glue::glue("pngtopnm {ref_png} > {ref_pnm}"))
 
 
-      # Because casting from double to unsigned char truncates rather than rounds,
-      # we get results which may be different by a magnitude of 1 (since {png}
-      # library rounds properly)
       dig1 <- readBin(new_pnm , what = 'raw', size = 1, n = 1e6)
       dig2 <- readBin(ref_pnm , what = 'raw', size = 1, n = 1e6)
 
-      expect_true(max(abs(as.integer(dig1) - as.integer(dig2))) <= 1)
-
-      # expect_identical(
-      #   digest::digest(readBin(new_pnm , what = 'raw', size = 1, n = 1e6), algo = 'sha1'),
-      #   digest::digest(readBin(ref_pnm , what = 'raw', size = 1, n = 1e6), algo = 'sha1')
-      # )
+      expect_identical(
+        digest::digest(readBin(new_pnm , what = 'raw', size = 1, n = 1e6), algo = 'sha1'),
+        digest::digest(readBin(ref_pnm , what = 'raw', size = 1, n = 1e6), algo = 'sha1')
+      )
 
     }
   }
@@ -86,8 +81,10 @@ test_that("write_png_core works for RGB", {
       dig1 <- readBin(new_pnm , what = 'raw', size = 1, n = 1e6)
       dig2 <- readBin(ref_pnm , what = 'raw', size = 1, n = 1e6)
 
-      expect_true(max(abs(as.integer(dig1) - as.integer(dig2))) <= 1)
-
+      expect_identical(
+        digest::digest(readBin(new_pnm , what = 'raw', size = 1, n = 1e6), algo = 'sha1'),
+        digest::digest(readBin(ref_pnm , what = 'raw', size = 1, n = 1e6), algo = 'sha1')
+      )
     }
   }